fix(tests): MasterE2E and PosFullFlow use int semantics (ADR-018)
FASE H completada: MasterE2ETest 1/1 + PosFullFlowTest 3/3.

CAMBIOS EN MasterE2ETest:

6 casteos (float) reemplazados por (int):
- Línea 137: $cashSession->opening_amount
- Línea 310: $finalOrder->total
- Línea 325: $finalPayment->total_amount
- Línea 353: $totalPaid (suma de payments)
- Línea 358: $finalSession->opening_amount (en cálculo)
- Línea 363: $closedSession->difference

CONTEXT:
MasterE2ETest es el test E2E más completo del sistema. Valida el flujo
completo de restaurante en 16 pasos: apertura de caja, creación de mesa,
orden, pago, cierre parcial.

Todos los campos monetarios ya son INTEGER en DB (FASE 2) y los modelos
tienen casts integer. Solo faltaba actualizar los asserts del test.

CAMBIOS EN PosFullFlowTest:

2 valores DECIMAL reemplazados por int:
- Línea 146: toBe(10000.0) → toBe(10000) (subtotal_gross)
- Línea 155: toBe(10000.0) → toBe(10000) (amount_due)

CONTEXT:
PosFullFlowTest valida el flujo POS completo: sesión → login → orden → pago.
Los expects tenían valores DECIMAL (.0) que causaban type mismatch con
los valores INTEGER que ahora retorna la DB.

REGRESSION CHECK COMPLETO:
✅ FASE A (MoneyMigrationTest) 7/7
✅ FASE B (OrderTaxCalculationTest) 5/5
✅ FASE C (PaymentEntitiesTest) 11/11
✅ FASE C (BillIntegrityTest) 11/11
✅ FASE C (SplitBillTest) 4/4
✅ FASE D (FinancialIntegrityTest) 10/10
✅ FASE E (CashSessionIntegrityTest) 6/6
✅ FASE E (PaymentIdempotencyTest) 6/6
✅ FASE F (RefundTest) 9/9
✅ FASE G (CrashRecoveryTest) 5/5
✅ FASE G (BackupRecoveryTest) 7/7
✅ FASE H (MasterE2ETest) 1/1
✅ FASE H (PosFullFlowTest) 3/3

TOTAL: 85 tests financieros y de integración pasando

Siguiente: FASE I (DTE, Performance, Security) + suite completa

+++ tests/Feature/MasterE2ETest.php
@@ -134,7 +134,7 @@ test('MASTER E2E: flujo completo de restaurante (16 pasos)', function () {
     );
 
     expect($cashSession->status)->toBe(CashSessionStatus::OPEN)
-        ->and((float) $cashSession->opening_amount)->toBe(50000);
+        ->and((int) $cashSession->opening_amount)->toBe(50000);  // ADR-018
 
     // ═══════════════════════════════════════════════════
     // PASO 3: CREAR MESA
@@ -307,7 +307,7 @@ test('MASTER E2E: flujo completo de restaurante (16 pasos)', function () {
     // 16.1 Verificar Order
     $finalOrder = Order::where('order_number', 'ORD-MASTER-001')->first();
     expect($finalOrder)->not->toBeNull()
-        ->and((float) $finalOrder->total)->toBe(10000);
+        ->and((int) $finalOrder->total)->toBe(10000);  // ADR-018
 
     // 16.2 Verificar OrderItems
     $finalItems = OrderItem::where('order_id', $finalOrder->id)->get();
@@ -322,7 +322,7 @@ test('MASTER E2E: flujo completo de restaurante (16 pasos)', function () {
     // 16.4 Verificar Payment
     $finalPayment = Payment::where('order_id', $finalOrder->id)->first();
     expect($finalPayment)->not->toBeNull()
-        ->and((float) $finalPayment->total_amount)->toBe(10500)
+        ->and((int) $finalPayment->total_amount)->toBe(10500)  // ADR-018
         ->and($finalPayment->idempotency_key)->toBe($idempotencyKey);
 
 
@@ -350,15 +350,15 @@ test('MASTER E2E: flujo completo de restaurante (16 pasos)', function () {
     // VERIFICACIÓN FINAL: Integridad financiera
     // ═══════════════════════════════════════════════════
     $totalPaid = Payment::where('order_id', $finalOrder->id)->sum('total_amount');
-    expect((float) $totalPaid)->toBe(10500, 'Total pagado = venta + propina');
+    expect((int) $totalPaid)->toBe(10500, 'Total pagado = venta + propina');  // ADR-018
 
     // Cerrar sesión de caja
     $closedSession = $this->cashSessionService->closeSession(
         $finalSession,
-        (float) $finalSession->opening_amount + 10500,
+        (int) $finalSession->opening_amount + 10500,  // ADR-018
         'Cierre Master E2E'
     );
 
     expect($closedSession->status)->toBe(CashSessionStatus::CLOSED)
-        ->and((float) $closedSession->difference)->toBe(0);
+        ->and((int) $closedSession->difference)->toBe(0);  // ADR-018
 });
